requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
sentence-transformers>=2.2.0
faiss-cpu>=1.8.0
langdetect>=1.0.9
deep-translator>=1.11.4
numpy>=1.21.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0